extract mock data
